+++ leader-election-impl/src/main/java/com/palantir/paxos/LeaderPinger.java
@@ -29,4 +29,5 @@ public interface LeaderPinger {
      */
     LeaderPingResult pingLeaderWithUuid(UUID uuid);
 
+    Boolean isCurrentAdjudicating();
 }
